requirements/requirements.txt
google-cloud-core==2.3.3
google-cloud-bigquery-storage ==2.23.0
google-cloud-storage==2.13.0
google-cloud-bigquery==3.13.0
google-cloud-secret-manager==2.16.0
google-api-python-client==2.108.0
google-api-core==2.14.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==1.5.3
pandas-gbq==0.19.2
numpy
db-dtypes==1.0.5
openpyxl==3.1.2
